LayoutTests/platform/win/css3/flexbox/flexitem-expected.txt


PASS .flexbox 1 
PASS .flexbox 2 
PASS .flexbox 3 
PASS .flexbox 4 
PASS .flexbox 5 
PASS .flexbox 6 
PASS .flexbox 7 
FAIL .flexbox 8 assert_equals: 
<div class="flexbox column" style="height:210px">
  <button data-expected-display="block" data-expected-width="600" data-expected-height="30">button</button>
  <canvas data-expected-display="block" data-expected-width="600" data-expected-height="30" style="height:30px">canvas</canvas>
  <iframe data-expected-display="block" data-expected-width="600" data-expected-height="30" style="height:30px" src="data:text/html,&lt;body bgcolor=#fff&gt;iframe&lt;/body&gt;"></iframe>
  <iframe seamless="" data-expected-display="block" data-expected-width="600" data-expected-height="30" style="height:30px" src="data:text/html,&lt;body bgcolor=#fff&gt;iframe&lt;/body&gt;"></iframe>
  <object data-expected-display="block" data-expected-width="600" data-expected-height="30">object</object>
  <select data-expected-display="block" data-expected-width="600" data-expected-height="30">
    <option>select</option>
  </select>
  <textarea data-expected-display="block" data-expected-width="600" data-expected-height="30">textarea</textarea>
</div>
height expected 30 but got 29
PASS .flexbox 9 
PASS .flexbox 10 
PASS .flexbox 11 
PASS .flexbox 12 
PASS .flexbox 13 
PASS .flexbox 14 
PASS .flexbox 15 
PASS .flexbox 16 
PASS .flexbox 17 
FAIL .flexbox 18 assert_equals: 
<div class="flexbox" style="height:200px">
  <button data-expected-display="block" data-expected-width="100" data-expected-height="200">button</button>
  <canvas data-expected-display="block" data-expected-width="100" data-expected-height="200">canvas</canvas>
  <iframe data-expected-display="block" data-expected-width="100" data-expected-height="200" src="data:text/html,&lt;body bgcolor=#fff&gt;iframe&lt;/body&gt;"></iframe>
  <object data-expected-display="block" data-expected-width="100" data-expected-height="200">object</object>
  <select data-expected-display="block" data-expected-width="100" data-expected-height="200">
    <option>select</option>
  </select>
  <textarea data-expected-display="block" data-expected-width="100" data-expected-height="200">textarea</textarea>
</div>
width expected 100 but got 96
button

object




button



button

object





Some text
Some more text
